Rename DataManagement/hooks.js to filterSortHooks
